support text styles for math equations.
