+++ src/FontSizeCommand.js
@@ -1,24 +1,17 @@
 // @flow
 
-import {Schema} from 'prosemirror-model';
-import {EditorState} from 'prosemirror-state';
-import {AllSelection, TextSelection} from 'prosemirror-state';
-import {Transform} from 'prosemirror-transform';
-import {EditorView} from 'prosemirror-view';
-
-import {MARK_FONT_SIZE} from './MarkNames';
-import applyMark from './applyMark';
 import UICommand from './ui/UICommand';
+import applyMark from './applyMark';
+import isTextStyleMarkCommandEnabled from './isTextStyleMarkCommandEnabled';
+import {EditorState} from 'prosemirror-state';
+import {EditorView} from 'prosemirror-view';
+import {MARK_FONT_SIZE} from './MarkNames';
+import {Schema} from 'prosemirror-model';
+import {Transform} from 'prosemirror-transform';
 
 function setFontSize(tr: Transform, schema: Schema, pt: number): Transform {
   const markType = schema.marks[MARK_FONT_SIZE];
   if (!markType) {
-    return tr;
-  }
-  const {selection} = tr;
-  if (
-    !(selection instanceof TextSelection || selection instanceof AllSelection)
-  ) {
     return tr;
   }
   const attrs = pt ? {pt} : null;
@@ -36,27 +29,7 @@ class FontSizeCommand extends UICommand {
   }
 
   isEnabled = (state: EditorState): boolean => {
-    const {schema, selection, tr} = state;
-    if (
-      !(selection instanceof TextSelection || selection instanceof AllSelection)
-    ) {
-      return false;
-    }
-    const markType = schema.marks[MARK_FONT_SIZE];
-    if (!markType) {
-      return false;
-    }
-
-    const {from, to} = selection;
-
-    if (to === from + 1) {
-      const node = tr.doc.nodeAt(from);
-      if (node.isAtom && !node.isText && node.isLeaf) {
-        // An atomic node (e.g. Image) is selected.
-        return false;
-      }
-    }
-    return true;
+    return isTextStyleMarkCommandEnabled(state, MARK_FONT_SIZE);
   };
 
   execute = (
