+++ src/isTextStyleMarkCommandEnabled.js
@@ -1,0 +1,42 @@
+// @flow
+
+import isNodeSelectionForNodeType from './isNodeSelectionForNodeType';
+import {EditorState} from 'prosemirror-state';
+import {MATH} from './NodeNames';
+import {AllSelection, TextSelection} from 'prosemirror-state';
+
+// Whether the command for apply specific text style mark is enabled.
+export default function isTextStyleMarkCommandEnabled(
+  state: EditorState,
+  markName: string
+): boolean {
+  const {selection, schema, tr} = state;
+  const markType = schema.marks[markName];
+  if (!markType) {
+    return false;
+  }
+  const mathNodeType = schema.nodes[MATH];
+  if (mathNodeType && isNodeSelectionForNodeType(selection, mathNodeType)) {
+    // A math node is selected.
+    return true;
+  }
+
+  if (
+    !(selection instanceof TextSelection || selection instanceof AllSelection)
+  ) {
+    // Could be a NodeSelection or CellSelection.
+    return false;
+  }
+
+  const {from, to} = state.selection;
+
+  if (to === from + 1) {
+    const node = tr.doc.nodeAt(from);
+    if (node.isAtom && !node.isText && node.isLeaf) {
+      // An atomic node (e.g. Image) is selected.
+      return false;
+    }
+  }
+
+  return true;
+}
